Add support for ERC1271
