+++ pkg/solidity-utils/contracts/test/SignaturesValidatorMock.sol
@@ -2,9 +2,9 @@
 
 pragma solidity ^0.7.0;
 
-import "../helpers/EOASignaturesValidator.sol";
+import "../helpers/SignaturesValidator.sol";
 
-contract EOASignaturesValidatorMock is EOASignaturesValidator {
+contract SignaturesValidatorMock is SignaturesValidator {
     event Authenticated();
 
     bytes32 public constant MOCK_AUTHENTICATE_TYPEHASH = keccak256("Authenticate(uint256 amount,uint256 nonce)");
@@ -13,14 +13,20 @@ contract EOASignaturesValidatorMock is EOASignaturesValidator {
         // solhint-disable-previous-line no-empty-blocks
     }
 
+    function getStructHash(uint256 amount, uint256 nonce) public pure returns (bytes32) {
+        return keccak256(abi.encode(MOCK_AUTHENTICATE_TYPEHASH, amount, nonce));
+    }
+
+    function getDigest(uint256 amount, uint256 nonce) public view returns (bytes32) {
+        return _hashTypedDataV4(getStructHash(amount, nonce));
+    }
+
     function authenticatedCall(
         address user,
         uint256 amount,
         bytes memory signature
     ) external {
-        bytes32 structHash = keccak256(abi.encode(MOCK_AUTHENTICATE_TYPEHASH, amount, getNextNonce(user)));
-
-        _ensureValidSignature(user, structHash, signature, Errors.INVALID_SIGNATURE);
+        _ensureValidSignature(user, getStructHash(amount, getNextNonce(user)), signature, Errors.INVALID_SIGNATURE);
         emit Authenticated();
     }
 
@@ -31,9 +37,13 @@ contract EOASignaturesValidatorMock is EOASignaturesValidator {
         uint256 deadline
     ) external {
         // Note that the deadline should be included in a real signing scheme - we keep it out here for simplicity.
-        bytes32 structHash = keccak256(abi.encode(MOCK_AUTHENTICATE_TYPEHASH, amount, getNextNonce(user)));
-
-        _ensureValidSignature(user, structHash, signature, deadline, Errors.INVALID_SIGNATURE);
+        _ensureValidSignature(
+            user,
+            getStructHash(amount, getNextNonce(user)),
+            signature,
+            deadline,
+            Errors.INVALID_SIGNATURE
+        );
         emit Authenticated();
     }
